refactor: router 변경 대응
